// Project/InternalAuditSystem.Web/ClientApp/src/app/swaggerapi/AngularFiles/model/aCMReportMappingAC.ts
/**
 * My Title
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ACMReviewerDocumentAC } from './aCMReviewerDocumentAC';
import { BaseModelAC } from './baseModelAC';
import { ReportUserType } from './reportUserType';
import { ReviewStatus } from './reviewStatus';

export interface ACMReportMappingAC extends BaseModelAC { 
    reportUserType: ReportUserType;
    userType?: string;
    acmId: string;
    userId: string;
    name?: string;
    designation?: string;
    status: ReviewStatus;
    statusName?: string;
    acmReviewerDocumentList?: Array<ACMReviewerDocumentAC>;
    isSelected: boolean;
    isAllowEdit: boolean;
    isAllowDelete: boolean;
    isAllowView: boolean;
    reportId: string;
    acmReportId?: string;
}