// Project/InternalAuditSystem.Web/ClientApp/src/app/swaggerapi/AngularFiles/model/aCMReportDetailAC.ts
/**
 * My Title
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ACMReportMappingAC } from './aCMReportMappingAC';
import { ACMReviewerAC } from './aCMReviewerAC';
import { BaseModelAC } from './baseModelAC';
import { EntityUserMappingAC } from './entityUserMappingAC';
import { KeyValuePairOfIntegerAndString } from './keyValuePairOfIntegerAndString';
import { ReportAC } from './reportAC';
import { ReportStage } from './reportStage';
import { ReportStatus } from './reportStatus';
import { ReviewStatus } from './reviewStatus';

export interface ACMReportDetailAC extends BaseModelAC { 
    entityId?: string;
    acmId?: string;
    acmReportTitle?: string;
    stageList?: Array<KeyValuePairOfIntegerAndString>;
    statusList?: Array<KeyValuePairOfIntegerAndString>;
    totalACMReport: number;
    auditorList?: Array<EntityUserMappingAC>;
    userList?: Array<EntityUserMappingAC>;
    acmReportReviewerList?: Array<ACMReviewerAC>;
    linkedACMReportList?: Array<ReportAC>;
    acmReportList?: Array<ACMReportMappingAC>;
    reviewerStatus: ReviewStatus;
    acmReportStatus: ReportStatus;
    acmReportStage: ReportStage;
    fileCount: number;
    acmReportStatusToString?: string;
    acmReportStageToString?: string;
    pageIndex: number;
}