// Project/InternalAuditSystem.Web/ClientApp/src/app/swaggerapi/AngularFiles/model/aCMReviewerAC.ts
/**
 * My Title
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ACMReviewerDocumentAC } from './aCMReviewerDocumentAC';
import { BaseModelAC } from './baseModelAC';
import { ReportUserType } from './reportUserType';
import { ReviewStatus } from './reviewStatus';

export interface ACMReviewerAC extends BaseModelAC { 
    acmPresentationId: string;
    userId: string;
    name?: string;
    designation?: string;
    status: ReviewStatus;
    statusName?: string;
    reportUserType: ReportUserType;
    userType?: string;
    reviewerDocumentList?: Array<ACMReviewerDocumentAC>;
}